Split EmbulkCommandLineParser into CliParser and CliParserImpl
