+++ embulk-core/src/main/java/org/embulk/deps/cli/CliParserImpl.java
@@ -12,8 +12,8 @@ import org.apache.commons.cli.Option;
 import org.apache.commons.cli.ParseException;
 import org.embulk.cli.EmbulkCommandLine;
 
-public class EmbulkCommandLineParser {
-    private EmbulkCommandLineParser(
+final class CliParserImpl extends CliParser {
+    CliParserImpl(
             final String usage,
             final List<AbstractHelpLineDefinition> helpLineDefinitions,
             final int minArgs,
@@ -37,69 +37,7 @@ public class EmbulkCommandLineParser {
         this.optionDefinitionFromCliOption = optionDefinitionFromCliOption;
     }
 
-    public static class Builder {
-        private Builder() {
-            this.mainUsage = null;
-            this.additionalUsage = new StringBuilder();
-            this.helpLineDefinitions = new ArrayList<AbstractHelpLineDefinition>();
-            this.minArgs = 0;
-            this.maxArgs = Integer.MAX_VALUE;
-            this.width = 74;
-        }
-
-        public EmbulkCommandLineParser build() {
-            return new EmbulkCommandLineParser(
-                    this.mainUsage + this.additionalUsage.toString(),
-                    this.helpLineDefinitions,
-                    this.minArgs,
-                    this.maxArgs,
-                    this.width);
-        }
-
-        public Builder setMainUsage(final String mainUsage) {
-            this.mainUsage = mainUsage;
-            return this;
-        }
-
-        public Builder addUsage(final String line) {
-            this.additionalUsage.append(System.getProperty("line.separator"));
-            this.additionalUsage.append(line);
-            return this;
-        }
-
-        public Builder addOptionDefinition(final OptionDefinition optionDefinition) {
-            this.helpLineDefinitions.add(optionDefinition);
-            return this;
-        }
-
-        public Builder addHelpMessageLine(final String message) {
-            this.helpLineDefinitions.add(new HelpMessageLineDefinition(message));
-            return this;
-        }
-
-        public Builder setArgumentsRange(final int minArgs, final int maxArgs) {
-            this.minArgs = minArgs;
-            this.maxArgs = maxArgs;
-            return this;
-        }
-
-        public Builder setWidth(final int width) {
-            this.width = width;
-            return this;
-        }
-
-        private String mainUsage;
-        private final StringBuilder additionalUsage;
-        private final ArrayList<AbstractHelpLineDefinition> helpLineDefinitions;
-        private int minArgs;
-        private int maxArgs;
-        private int width;
-    }
-
-    public static Builder builder() {
-        return new Builder();
-    }
-
+    @Override
     public EmbulkCommandLine parse(final List<String> argsEmbulk,
                                    final List<String> jrubyOptions,
                                    final PrintWriter helpPrintWriter,
@@ -144,10 +82,7 @@ public class EmbulkCommandLineParser {
         return commandLineBuilder.build();
     }
 
-    public final void printHelp(final PrintStream printStream) {
-        this.printHelp(new PrintWriter(printStream));
-    }
-
+    @Override
     public void printHelp(final PrintWriter printWriter) {
         final CliHelpFormatterWithHelpMessages helpFormatter = new CliHelpFormatterWithHelpMessages("Usage: ", 32);
         helpFormatter.printHelp(
